requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.28.63
botocore==1.31.63
cachetools==5.3.0
certifi==2022.12.7
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.5
contourpy==1.1.1
cycler==0.10.0
decorator==4.4.2
Django==5.0.3
django-cors-headers==4.2.0
django-nextjs==2.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.4
fire==0.5.0
firebase==3.0.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.3.3
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.4.0
gcloud==0.17.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.53.0
h5py==3.8.0
httplib2==0.20.4
huggingface-hub==0.18.0
idna==2.10
imageio==2.20.0
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jws==0.1.3
keras==2.12.0
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.0
ml-dtypes==0.0.4
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.0
networkx==3.1
numpy==1.25.2
oauth2client==3.0.0
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.3
Pillow==10.0.1
proglog==0.1.10
protobuf==4.22.1
psutil==5.9.5
psycopg2==2.9.6
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycryptodome==3.4.3
PyJWT==2.8.0
pyotp==2.8.0
pyparsing==2.4.7
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-firebase==1.2
python-jwt==2.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
roboflow==1.1.7
rsa==4.9
s3transfer==0.7.0
sahi==0.11.14
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
supervision==0.15.0
sympy==1.12
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.64.0
typing_extensions==4.7.1
tzdata==2024.1
ultralytics==8.0.197
urllib3==2.0.4
websockets==11.0.3
Werkzeug==2.3.7
whitenoise==6.6.0
wrapt==1.14.1
yarl==1.9.2
yolov5==7.0.12
